refactor: Split test file into module-based test files
- Split Tyme4SwiftTests.swift (2030 lines, 114 tests) into 10 module-based files
- Organized tests by domain: Solar, Lunar, SixtyCycle, EightChar, Culture, God, Fetus, JulianDay, Regression, EquatableHashable
- All 114 tests pass with no failures
- Regression tests (Issue #14, #19, #42) grouped in dedicated RegressionTests.swift
- Removed original monolithic test file to reduce merge conflicts

Part of Issue #33
