--- Tests/tymeTests/tyme4swiftTests.swift
@@ -1,6 +1,0 @@
-import Testing
-@testable import tyme
-
-@Test func example() async throws {
-    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-}
